+++ NEWS.md
@@ -7,6 +7,9 @@
   case insensitive when reading and writing, so forcing lower case removes any
   potential ambiguity.
 
+* Character limit checks in `write_xpt()` now check for the number of bytes
+  instead of the number of characters (#746).
+
 # haven 2.5.5
 
 * Updated ReadStat to fix stricter gcc diagnostics.